Modification SimPheno
Modification of the function to simulate the phenotype by changing the way allelic effects are sampled. Now they are sampled from a uniform distribution (1-10). The sign of the effect is randomly sampled.

I also extended the SimPheno function to collect the number of non zero effect per QTLs and the number of QTL effects (alleles)