Modif 2: modif bi-allelic QTL effect
I modified the way the QTL effect is attributed for the bi-allelic QTL. Before, the QTL effect was shared between the major and the minor allele. The major allele was always negative and the minor allele positive. Now only the minor allele get an effect that can be positive or negative.

I went back to the situation where I peacked the largest ancestral class for the QTL 6.